backend/requirements.txt
fastapi==0.136.1
uvicorn==0.46.0
psycopg2-binary==2.9.12
pydantic==2.13.3
pydantic-settings==2.14.0
pdfplumber==0.11.9
python-docx==1.2.0
langchain==1.2.15
langchain-text-splitters==1.1.2
huggingface-hub==1.12.0
google-generativeai==0.8.3
python-multipart==0.0.27
pgvector==0.4.2
sqlalchemy==2.0.49
alembic==1.18.4
